// milo/ui/src/proto/turboci/graph/orchestrator/v1/value_ref.pb.ts
export interface ValueRef {
  /**
   * The Type URL of the referenced data.
   *
   * Will have the form `type.googleapis.com/proto.package.MessageName`.
   *
   * Note that for inline data, this is duplicated with data.inline.*.type_url.
   */
  readonly typeUrl?:
    | string
    | undefined;
  /**
   * The concrete realm to which this data belongs.
   *
   * Note that this will always be set, and will never be "$from_token" or
   * "$from_container" as you may write via [ValueWrite.realm].
   */
  readonly realm?:
    | string
    | undefined;
  /**
   * Serialized [ValueDigest] for the data.
   *
   * This indicates that the data must be retrieved from the `value_data` map
   * in the response (or from some other CAS system using `digest` as the
   * effective key).
   */
  readonly digest?:
    | string
    | undefined;
  /**
   * Inline contains the actual data, if the orchestrator decided that this
   * data was too small/unique to store indirectly via `digest`.
   *
   * It is expected that clients will always use a library function to handle
   * inline vs. digest-referenced data. Server-provided JSON via ValueData
   * will always be provided via digest, rather than stored inline.
   *
   * See [ValueDigest] for how to compute `digest` from `inline`, should
   * it be necessary.
   */
  readonly inline?:
    | Any
    | undefined;
  /**
   * If set, the reason this ValueRef's content was omitted from the response.
   *
   * If unset, it means that the caller has access to this data, and that the
   * data matched the [ValueFilter] provided by the user.
   *
   * NOTE: For paged responses it is possible that the actual [ValueData] will
   * be provided in a subsequent page. In this case, OmitReason will be unset,
   * even if the [ValueData] is not in the current page.
   */
  readonly omitReason?: OmitReason | undefined;
}

function createBaseValueRef(): ValueRef {
  return { typeUrl: undefined, realm: undefined, digest: undefined, inline: undefined, omitReason: undefined };
}

export const ValueRef: MessageFns<ValueRef> = {
  encode(message: ValueRef, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.typeUrl !== undefined) {
      writer.uint32(10).string(message.typeUrl);
    }
    if (message.realm !== undefined) {
      writer.uint32(18).string(message.realm);
    }
    if (message.digest !== undefined) {
      writer.uint32(26).string(message.digest);
    }
    if (message.inline !== undefined) {
      Any.encode(message.inline, writer.uint32(34).fork()).join();
    }
    if (message.omitReason !== undefined) {
      writer.uint32(40).int32(message.omitReason);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ValueRef {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseValueRef() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.typeUrl = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.realm = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.digest = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.inline = Any.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.omitReason = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ValueRef {
    return {
      typeUrl: isSet(object.typeUrl) ? globalThis.String(object.typeUrl) : undefined,
      realm: isSet(object.realm) ? globalThis.String(object.realm) : undefined,
      digest: isSet(object.digest) ? globalThis.String(object.digest) : undefined,
      inline: isSet(object.inline) ? Any.fromJSON(object.inline) : undefined,
      omitReason: isSet(object.omitReason) ? omitReasonFromJSON(object.omitReason) : undefined,
    };
  },

  toJSON(message: ValueRef): unknown {
    const obj: any = {};
    if (message.typeUrl !== undefined) {
      obj.typeUrl = message.typeUrl;
    }
    if (message.realm !== undefined) {
      obj.realm = message.realm;
    }
    if (message.digest !== undefined) {
      obj.digest = message.digest;
    }
    if (message.inline !== undefined) {
      obj.inline = Any.toJSON(message.inline);
    }
    if (message.omitReason !== undefined) {
      obj.omitReason = omitReasonToJSON(message.omitReason);
    }
    return obj;
  },

  create(base?: DeepPartial<ValueRef>): ValueRef {
    return ValueRef.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ValueRef>): ValueRef {
    const message = createBaseValueRef() as any;
    message.typeUrl = object.typeUrl ?? undefined;
    message.realm = object.realm ?? undefined;
    message.digest = object.digest ?? undefined;
    message.inline = (object.inline !== undefined && object.inline !== null)
      ? Any.fromPartial(object.inline)
      : undefined;
    message.omitReason = object.omitReason ?? undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}